pyproject.toml
[tool.poetry]
name = "aituringbackend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
psycopg2 = "^2.9.9"
dj-database-url = "^2.1.0"
python-decouple = "^3.8"
django-environ = "^0.11.2"
whitenoise = "6.6.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
asgiref = "3.7.2"
python-dotenv = "1.0.1"
pytz = "2024.1"
typing-extensions = "4.9.0"
tzdata = "2024.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
django-debug-toolbar = "^4.3.0"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 90
target-version = "py310"
indent-width = 4
show-fixes = true

[tool.ruff.lint]
# # Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.lint.isort]
default-section = "third-party"
no-lines-before = ["future", "standard-library"]
forced-separate = ["tests"]
force-wrap-aliases = true
section-order = [
  "future",
  "standard-library",
  "django",
  "third-party",
  "first-party",
  "local-folder",
]

[tool.ruff.lint.isort.sections]
"django" = ["django"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
